BIC {stats4}	R Documentation

Bayesian Information Criterion

Description

This generic function calculates the Bayesian information criterion, also known as Schwarz's Bayesian criterion (SBC), for one or several fitted model objects for which a log-likelihood value can be obtained, according to the formula -2*log-likelihood + npar*log(nobs), where npar represents the number of parameters and nobs the number of observations in the fitted model.

Usage

BIC(object, ...)

Arguments

object 	 An object of a suitable class for the BIC to be calculated - usually a "logLik" object or an object for which a logLik method exists.

... 	 Some methods for this generic function may take additional, optional arguments. At present none do.

Value

Returns a numeric value with the corresponding BIC.

References

Schwarz, G. (1978) "Estimating the Dimension of a Model", Annals of Statistics, 6, 461-464.

See Also

logLik-methods, AIC-methods

Examples

lm1 <- lm(Fertility ~ . , data = swiss)
AIC(lm1)
BIC(lm1)

[Package stats4 version 2.9.2 Index]