mle-class {stats4}	R Documentation

Class "mle" for Results of Maximum Likelihood Estimation

Description

This class encapsulates results of a generic maximum likelihood procedure.

Objects from the Class

Objects can be created by calls of the form new("mle", ...), but most often as the result of a call to mle.

Slots

call:

Object of class "language". The call to mle.

coef:

Object of class "numeric". Estimated parameters.

fullcoef:

Object of class "numeric". Fixed and estimated parameters.

vcov:

Object of class "matrix". Approximate variance-covariance matrix.

min:

Object of class "numeric". Minimum value of objective function.

details:

a "list", as returned from optim.

minuslogl:

Object of class "function". The negative loglikelihood function.

nobs:

"integer" of length one. The number of observations (often NA, when not set in call explicitly).

method:

Object of class "character". The optimization method used.

Methods

confint

signature(object = "mle"): Confidence intervals from likelihood profiles.

logLik

signature(object = "mle"): Extract maximized log-likelihood.

profile

signature(fitted = "mle"): Likelihood profile generation.

nobs

signature(object = "mle"): Number of observations, here simply accessing the nobs slot mentioned above.

show

signature(object = "mle"): Display object briefly.

summary

signature(object = "mle"): Generate object summary.

update

signature(object = "mle"): Update fit.

vcov

signature(object = "mle"): Extract variance-covariance matrix.

[Package stats4 version 3.6.1 Index]