plot-methods {stats4}	R Documentation

Methods for Function plot in Package stats4

Description

Plot profile likelihoods for "mle" objects.

Usage


## S4 method for signature 'profile.mle,missing'
plot(x, levels, conf = c(99, 95, 90, 80, 50)/100, nseg = 50,
     absVal = TRUE, ...)

Arguments

x
an object of class "profile.mle"

levels
levels, on the scale of the absolute value of a t statistic, at which to interpolate intervals. Usually conf is used instead of giving levels explicitly.

conf
a numeric vector of confidence levels for profile-based confidence intervals on the parameters.

nseg
an integer value giving the number of segments to use in the spline interpolation of the profile t curves.

absVal
a logical value indicating whether or not the plots should be on the scale of the absolute value of the profile t. Defaults to TRUE.

...
other arguments to the plot function can be passed here.

Methods

signature(x = "ANY", y = "ANY")

Generic function: see plot.

signature(x = "profile.mle", y = "missing")

Plot likelihood profiles for x.

[Package stats4 version 3.6.1 Index]