update-methods {stats4}	R Documentation

Methods for Function update in Package stats4

Description

Update "mle" objects.

Usage


## S4 method for signature 'mle'
update(object, ..., evaluate = TRUE)

Arguments

object
An existing fit.

...
Additional arguments to the call, or arguments with changed values. Use name = NULL to remove the argument name.

evaluate
If true evaluate the new call else return the call.

Methods

signature(object = "ANY")

Generic function: see update.

signature(object = "mle")

Update a fit.

Examples


x <- 0:10
y <- c(26, 17, 13, 12, 20, 5, 9, 8, 5, 4, 8)
ll <- function(ymax = 15, xhalf = 6)
    -sum(stats::dpois(y, lambda = ymax/(1+x/xhalf), log = TRUE))
fit <- mle(ll)
## note the recorded call contains ..1, a problem with S4 dispatch
update(fit, fixed = list(xhalf = 3))

[Package stats4 version 3.6.1 Index]