profile-methods {stats4}	R Documentation

Methods for Function profile in Package ‘stats4’

Description

Profile likelihood for "mle" objects.

Usage

## S4 method for signature 'mle':
profile(fitted, which = 1:p, maxsteps = 100, alpha = 0.01,
        zmax = sqrt(qchisq(1 - alpha, 1L)), del = zmax/5,
        trace = FALSE, ...)

Arguments

fitted 	 Object to be profiled

which 	 Optionally select subset of parameters to profile.

maxsteps 	 Maximum number of steps to bracket zmax.

alpha 	 Significance level corresponding to zmax, based on a Scheffe-style multiple testing interval. Ignored if zmax is specified.

zmax 	 Cutoff for the profiled value of the signed root-likelihood.

del 	 Initial stepsize on root-likelihood scale.

trace 	 Logical. Print intermediate results.

... 	 Currently unused.

Details

The profiling algoritm tries to find an approximately evenly spaced set of at least five parameter values (in each direction from the optimum) to cover the root-likelihood function. Some care is taken to try and get sensible results in cases of high parameter curvature. Notice that it may not always be possible to obtain the cutoff value, since the likelihood might level off.

Value

An object of class "profile.mle", see "profile.mle-class".

Methods

fitted = "ANY"

Generic function: see profile.

fitted = "mle"

Profile the likelihood in the vicinity of the optimum of an "mle" object.

[Package stats4 version 2.9.2 Index]