profile.mle-class {stats4}	R Documentation

Class "profile.mle"; Profiling information for "mle" object

Description

Likelihood profiles along each parameter of likelihood function

Objects from the Class

Objects can be created by calls of the form new("profile.mle", ...), but most often by invoking profile on an "mle" object.

Slots

profile:

Object of class "list". List of profiles, one for each requested parameter. Each profile is a data frame with the first column called z being the signed square root of the -2 log likelihood ratio, and the others being the parameters with names prefixed by par.vals.

summary:

Object of class "summary.mle". Summary of object being profiled.

Methods

confint

signature(object = "profile.mle"): Use profile to generate approximate confidence intervals for parameters.

plot

signature(x = "profile.mle", y = "missing"): Plot profiles for each parameter.

See Also

mle, mle-class, summary.mle-class

[Package stats4 version 2.9.2 Index]