summary.mle-class {stats4}	R Documentation

Class "summary.mle", summary of "mle" objects

Description

Extract of "mle" object

Objects from the Class

Objects can be created by calls of the form new("summary.mle", ...), but most often by invoking summary on an "mle" object. They contain values meant for printing by show.

Slots

call:

Object of class "language" The call that generated the "mle" object.

coef:

Object of class "matrix". Estimated coefficients and standard errors

m2logL:

Object of class "numeric". Minus twice the log likelihood.

Methods

show

signature(object = "summary.mle"): Pretty-prints object

coef

signature(object = "summary.mle"): Extracts the contents of the coef slot

See Also

summary, mle, mle-class

[Package stats4 version 2.9.2 Index]